+++ src/main/java/benchmarks/ants/colonies/colony/AntsColony.java
@@ -16,7 +16,7 @@
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package benchmarks.ants.colony;
+package benchmarks.ants.colonies.colony;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -31,10 +31,10 @@ import javax.annotation.Nonnull;
 import javax.annotation.concurrent.ThreadSafe;
 
 import benchmarks.ants.colonies.AntsSettings;
-import benchmarks.ants.colony.ant.AntRunResult;
-import benchmarks.ants.colony.ant.RunningAnt;
+import benchmarks.ants.colonies.colony.ant.AntRunResult;
+import benchmarks.ants.colonies.colony.ant.RunningAnt;
+import benchmarks.ants.colonies.parallelisation.ContinuousParallelExecutor;
 import benchmarks.ants.data.IDistancesData;
-import benchmarks.ants.parallelisation.ContinuousParallelExecutor;
 import benchmarks.metrics.PerformanceMeasurer;
 import benchmarks.metrics.PerformanceMeasuresCompiler;
 import util.Restrictions;
