+++ src/main/java/benchmarks/ants/colonies/colony/AntsStatistics.java
@@ -16,7 +16,7 @@
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package benchmarks.ants.colony;
+package benchmarks.ants.colonies.colony;
 
 import com.sun.istack.internal.NotNull;
 
@@ -34,7 +34,7 @@ import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
 import javax.annotation.concurrent.ThreadSafe;
 
-import benchmarks.ants.colony.ant.AntRunResult;
+import benchmarks.ants.colonies.colony.ant.AntRunResult;
 
 /**
  * Package local statistics for AntsColony work.
