+++ src/main/java/benchmarks/ants/colonies/colony/ant/RouteFinder.java
@@ -16,7 +16,7 @@
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package benchmarks.ants.colony.ant;
+package benchmarks.ants.colonies.colony.ant;
 
 import java.util.Optional;
 import java.util.concurrent.ThreadLocalRandom;
@@ -25,7 +25,7 @@ import javax.annotation.Nonnegative;
 import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
 
-import benchmarks.ants.colony.CachedRawEdgeQualities;
+import benchmarks.ants.colonies.colony.CachedRawEdgeQualities;
 import benchmarks.ants.data.IDistancesData;
 
 /**
