+++ src/main/java/benchmarks/ants/colonies/parallelisation/AgentsThreadsFactory.java
@@ -16,7 +16,7 @@
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package benchmarks.ants.parallelisation;
+package benchmarks.ants.colonies.parallelisation;
 
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.atomic.AtomicInteger;
