+++ src/main/java/benchmarks/ants/colonies/parallelisation/ParallelBarrierExecutor.java
@@ -16,7 +16,7 @@
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package benchmarks.ants.parallelisation;
+package benchmarks.ants.colonies.parallelisation;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
